Replace GameEnginePropertyReader with ProductConfiguration
