--- .travis/update_engine_version
@@ -1,13 +1,0 @@
-#!/bin/bash
-#
-# This script updates the engine version compiled into the code.
-#
-
-readonly ENGINE_VERSION=$1
-readonly GAME_ENGINE_PROPERTIES=game-core/src/main/resources/META-INF/triplea/game_engine.properties
-
-echo "Updating engine version to '$ENGINE_VERSION'"
-
-# Update engine property file in-place, append build number.
-# e.g.: "engine_version = 1.9.0.0" -> "engine_version = 1.9.0.0.1234"
-sed -i "s/^\(engine_version\s*=\).*/\1 ${ENGINE_VERSION}/" "$GAME_ENGINE_PROPERTIES"
