+++ game-core/src/main/java/games/strategy/engine/ClientContext.java
@@ -2,7 +2,8 @@ package games.strategy.engine;
 
 import java.util.List;
 
-import games.strategy.engine.config.client.GameEnginePropertyReader;
+import org.triplea.common.config.product.ProductConfiguration;
+
 import games.strategy.engine.framework.map.download.DownloadCoordinator;
 import games.strategy.engine.framework.map.download.DownloadFileDescription;
 import games.strategy.engine.framework.map.download.DownloadRunnable;
@@ -45,7 +46,7 @@ import games.strategy.util.Version;
 public final class ClientContext {
   private static final ClientContext instance = new ClientContext();
 
-  private final GameEnginePropertyReader gameEnginePropertyReader = new GameEnginePropertyReader();
+  private final ProductConfiguration productConfiguration = new ProductConfiguration();
   private final MapDownloadController mapDownloadController = new MapDownloadController();
   private final DownloadCoordinator downloadCoordinator = new DownloadCoordinator();
 
@@ -60,7 +61,7 @@ public final class ClientContext {
   }
 
   public static Version engineVersion() {
-    return instance.gameEnginePropertyReader.getEngineVersion();
+    return instance.productConfiguration.getVersion();
   }
 
   public static List<DownloadFileDescription> getMapDownloadList() {
