+++ game-core/src/main/java/org/triplea/common/config/PropertyReader.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.config;
+package org.triplea.common.config;
 
 import static com.google.common.base.Preconditions.checkNotNull;
