--- game-core/src/main/resources/META-INF/triplea/game_engine.properties
@@ -1,1 +1,0 @@
-engine_version = 1.9.0.0.dev
