+++ game-core/src/main/resources/META-INF/triplea/product.properties
@@ -1,0 +1,1 @@
+version = 1.9.0.0.@qualifier@
