+++ game-core/src/test/resources/org/triplea/common/config/ResourcePropertyReaderTest.properties
