+++ game-headed/build.gradle
@@ -8,7 +8,6 @@ plugins {
 archivesBaseName = "$group-$name"
 description = 'TripleA Headed Game Client'
 mainClassName = 'org.triplea.game.headed.runner.HeadedGameRunner'
-version = getEngineVersion()
 
 ext {
     javaFxRuntimeUrl = 'https://github.com/triplea-game/assets/raw/master/javafx/jfxrt-1.8.0_181.jar'
