+++ lobby/src/test/java/org/triplea/lobby/server/login/LobbyLoginValidatorTest.java
@@ -24,6 +24,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mindrot.jbcrypt.BCrypt;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.triplea.common.config.MemoryPropertyReader;
 import org.triplea.lobby.common.LobbyConstants;
 import org.triplea.lobby.common.login.LobbyLoginChallengeKeys;
 import org.triplea.lobby.common.login.LobbyLoginResponseKeys;
@@ -40,7 +41,6 @@ import org.triplea.test.common.security.TestSecurityUtils;
 
 import com.google.common.collect.ImmutableMap;
 
-import games.strategy.engine.config.MemoryPropertyReader;
 import games.strategy.engine.lobby.server.userDB.DBUser;
 import games.strategy.util.Md5Crypt;
 import games.strategy.util.Tuple;
